encoder/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pysrt>=1.1.2
webvtt-py>=0.4.6
lz4>=4.3.0